third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/disconnected-distribution-cycle-expected.txt

CONSOLE ERROR: line 15: Uncaught TypeError: host.createShadowRoot is not a function
Distributions from disconnected subtrees should be cleared when inserted again to avoid cycles.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
